Внёс мелкие правки в Entity, для обеспечения создания сущностей в БД.…

+++ src/main/java/pro/sky/graduate_work_group5_team1/Repositories/CommentRepository.java
@@ -3,7 +3,9 @@ package pro.sky.graduate_work_group5_team1.Repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import pro.sky.graduate_work_group5_team1.entity.AdsComment;
+import pro.sky.graduate_work_group5_team1.model.AdsComment;
+
 @Repository
-public interface CommentRepository extends JpaRepository<AdsComment, Long> {
+public interface CommentRepository extends JpaRepository<AdsComment, Integer> {
+
 }
